Only audit "download" API events
After discussion with Joseph and the team it was decided that in order
to comply with ESA requirements we need to record which user requested
a download (and therefore the area of research that they nominated).
However, for other API events we can either glean enough information
from the Apache logs or we don't care about them.